+++ CHANGES.md
@@ -11,6 +11,8 @@ The released versions correspond to PyPI releases.
   compatible, we cannot exclude that we missed some problems.
 * Under macOS, at test start a symlink `/tmp` to the actual temporary directory is
   now created in the fake filesystem.
+* Patching of parsers for pandas >= 1.2 is removed since pandas now uses Python fs functions
+  internally even when the engine selected is "c".
 
 ### Features
 * added possibility to set a path inaccessible under Windows by using `chown()` with